refactor(core): split manifest module responsibilities
